latex: Move ExtBabel to sphinx.builders.latex.util package
